Refactor test strcuture to PSR-4
* Added PSR-4 mapping in composer
* Updated namespaces in tests
* Rename any files that do not conform to PSR-4
* Updated tests to "use" classes instead of requiring those
