+++ test/EnumFieldTest.php
@@ -1,10 +1,13 @@
 <?php
-namespace YouTrack;
+namespace YouTrack\Test;
+
+use YouTrack\EnumBundle;
+use YouTrack\EnumField;
 
 class EnumFieldTest extends \PHPUnit_Framework_TestCase
 {
 
-    private $filename = 'test/testdata/enum_bundle.xml';
+    private $filename = __DIR__  . '/testdata/enum_bundle.xml';
 
     public function testGetOwnedFieldBundleElements()
     {
