+++ test/IssueTest.php
@@ -1,6 +1,8 @@
 <?php
-namespace YouTrack;
-require_once 'requirements.php';
+namespace YouTrack\Test;
+
+
+use YouTrack\Issue;
 
 /**
  * Unit test for the youtrack issue class.
@@ -11,7 +13,7 @@ require_once 'requirements.php';
 class IssueTest extends \PHPUnit_Framework_TestCase
 {
 
-    private $filename = 'test/testdata/issue.xml';
+    private $filename = __DIR__  . '/testdata/issue.xml';
 
     public function testConstruct01()
     {
