--- test/requirements.php
@@ -1,6 +1,0 @@
-<?php
-/**
- * Include the requirements for testing.
- */
-
-require_once './vendor/autoload.php';
